requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.17.1
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.2
pytest-django==4.12.0
sqlparse==0.5.5
